Scripts/Misc: Migrate Operation Gnomeregan & Zalazane's Fall scripts to own files
