+++ src/server/scripts/Events/events_script_loader.cpp
@@ -18,6 +18,8 @@
 // This is where scripts' loading functions should be declared:
 void AddSC_event_childrens_week();
 void AddSC_event_fireworks();
+void AddSC_event_operation_gnomeregan();
+void AddSC_event_zalazane_fall();
 
 // The name of this function should match:
 // void Add${NameOfDirectory}Scripts()
@@ -25,4 +27,6 @@ void AddEventsScripts()
 {
     AddSC_event_childrens_week();
     AddSC_event_fireworks();
+    AddSC_event_operation_gnomeregan();
+    AddSC_event_zalazane_fall();
 }
