+++ src/server/scripts/Spells/spell_quest.cpp
@@ -1774,57 +1774,6 @@ class spell_q12308_escape_from_silverbrook_summon_worgen : public SpellScript
     }
 };
 
-enum BasicOrdersEmote
-{
-    SPELL_TEST_SALUTE        = 73835,
-    SPELL_TEST_ROAR          = 73836,
-    SPELL_TEST_CHEER         = 73725,
-    SPELL_TEST_DANCE         = 73837,
-    SPELL_TEST_STOP_DANCE    = 73886
-};
-
-/* 73725 - [DND] Test Cheer
-   73835 - [DND] Test Salute
-   73836 - [DND] Test Roar
-   73837 - [DND] Test Dance
-   73886 - [DND] Test Stop Dance */
-class spell_q25199_emote : public AuraScript
-{
-    PrepareAuraScript(spell_q25199_emote);
-
-    void HandlePeriodic(AuraEffect const* /*aurEff*/)
-    {
-        Unit* target = GetTarget();
-
-        switch (GetSpellInfo()->Id)
-        {
-            case SPELL_TEST_SALUTE:
-                target->HandleEmoteCommand(EMOTE_ONESHOT_SALUTE);
-                break;
-            case SPELL_TEST_ROAR:
-                target->HandleEmoteCommand(EMOTE_ONESHOT_ROAR);
-                break;
-            case SPELL_TEST_CHEER:
-                target->HandleEmoteCommand(EMOTE_ONESHOT_CHEER);
-                break;
-            case SPELL_TEST_DANCE:
-                target->SetUInt32Value(UNIT_NPC_EMOTESTATE, EMOTE_STATE_DANCE);
-                break;
-            case SPELL_TEST_STOP_DANCE:
-                target->SetUInt32Value(UNIT_NPC_EMOTESTATE, EMOTE_STATE_NONE);
-                break;
-            default:
-                return;
-        }
-        Remove();
-    }
-
-    void Register() override
-    {
-        OnEffectPeriodic += AuraEffectPeriodicFn(spell_q25199_emote::HandlePeriodic, EFFECT_0, SPELL_AURA_PERIODIC_DUMMY);
-    }
-};
-
 enum DeathComesFromOnHigh
 {
     SPELL_FORGE_CREDIT                  = 51974,
@@ -2436,7 +2385,6 @@ void AddSC_quest_spell_scripts()
     RegisterSpellScript(spell_q11896_weakness_to_lightning_46444);
     RegisterSpellScript(spell_q12308_escape_from_silverbrook_summon_worgen);
     RegisterSpellScript(spell_q12308_escape_from_silverbrook);
-    RegisterSpellScript(spell_q25199_emote);
     RegisterSpellScript(spell_q12641_death_comes_from_on_high);
     RegisterSpellScript(spell_q12641_recall_eye_of_acherus);
     RegisterSpellScript(spell_q12619_emblazon_runeblade);
